fixing register method

+++ controlers/userController.js
@@ -24,11 +24,16 @@ module.exports.getUser = async (req, res) => {
 };
 
 module.exports.deleteUser = async (req, res) => {
-  console.log(req.body.id);
-  await User.deleteOne({ _id: req.body.id });
+  // eslint-disable-next-line no-underscore-dangle
+  if (req.role === 'admin' || req.params._id === req.user_id) {
+    await User.deleteOne({ _id: req.body.id });
 
-  return res.status(200).json({
-    status: 'success',
+    return res.status(200).json({
+      status: 'success',
+    });
+  }
+  return res.status(403).json({
+    message: 'access denied',
   });
 };
 
@@ -40,7 +45,7 @@ module.exports.resetPass = async (req, res) => {
   }
 
   // eslint-disable-next-line no-underscore-dangle
-  if (req.role === 'Admin' || req.user_id === user._id) {
+  if (req.role === 'Admin' || req.user_id === req.params._id) {
     if (req.body.password === req.body.passwordConfirm) {
       const encryptedPassword = await bcrypt.hash(req.body.password, 10);
       user.password = encryptedPassword;
